+++ client/src/comp/WriteModal.js
@@ -109,7 +109,8 @@ function WriteModal({
   accToken,
   setAccToken,
   userInfo,
-  setisLogin,
+  setNeedUpdate,
+  needUpdate,
 }) {
   const url = process.env.REACT_APP_API_URL || `http://localhost:4000`;
   const history = useHistory();
@@ -143,8 +144,8 @@ function WriteModal({
           setAccToken(writeRes.data.accessToken);
         }
         alert("새로운 줄임말이 등록되었습니다.");
+        setNeedUpdate(!needUpdate);
         setWriteModal(false);
-        history.push("/mypage");
       }
     } catch (error) {
       alert("다시 로그인해주세요.");
